pyproject.toml
[project]
name = "remnawave-ansible-gen"
version = "0.1.0"
description = "Generator for Remnawave Ansible modules"
requires-python = ">=3.11"
dependencies = [
    "prance>=23.6.0",
    "openapi-spec-validator>=0.7.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
generate = "remnawave_ansible_gen.cli:main"
version = "remnawave_ansible_gen.version:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/remnawave_ansible_gen"]

[dependency-groups]
dev = [
    "ansible>=2.14.0",
    "ansible-lint>=24.0.0",
    "molecule>=6.0.0",
    "molecule-docker>=2.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"ansible_collections/**/*.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
extra_checks = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
files = ["src/remnawave_ansible_gen"]

[[tool.mypy.overrides]]
module = [
    "prance",
    "prance.*",
    "openapi_spec_validator",
    "openapi_spec_validator.*",
    "jinja2",
    "jinja2.*",
    "yaml",
    "yaml.*",
]
ignore_missing_imports = true
